requirements.txt
gradio>=4.44.0
langgraph>=0.2.3
langchain>=0.3
httpx>=0.28.1
python-dotenv>=1.1.1
pytest>=8.4.1
black>=25.1.0
isort>=6.0.1
ruff>=0.12.10
mcp
langchain-mcp-adapters
langchain-mistralai